Refactor pbench-cli
Consolidate pbench-cli to make it easier to backport commands
so that the workload scripts can still keep on working when
we convert more bash scripts to python.
