+++ pbench/cli/agent/commands/metalog.py
@@ -26,21 +26,3 @@ def metalog_add_option(logfile, section, option, value):
     where the iterations are in the <iterations.file>, one iteration per line
     """
     add_metalog_option(logfile, section, option, value)
-
-
-#
-# pbench backwards compatible
-#
-@click.command()
-@click.argument("logfile")
-@click.argument("section")
-@click.argument("option")
-@click.argument("value")
-def metalog_add(logfile, section, option, value):
-    """Add an option to a section of the metadata.log file.
-
-    E.g. using an 'iterations' arg for the option
-    iterations: 1-iter, 2-iter, 3-iter
-    where the iterations are in the <iterations.file>, one iteration per line
-    """
-    add_metalog_option(logfile, section, option, value)
