+++ pbench/cli/agent/shell.py
@@ -1,7 +1,51 @@
 import click
 
-from pbench.cli.agent.commands import metalog
-from pbench.cli.agent.commands import results
+from pbench.cli.agent import options
+from pbench.cli.agent.commands.log import add_metalog_option
+from pbench.cli.agent.commands.results import move_results
+
+
+#
+# log subcommand
+#
+@click.group()
+@click.pass_context
+def metalog(ctx):
+    """Used for passing the context from the main cli"""
+    pass
+
+
+@metalog.group()
+@click.argument("logfile")
+@click.argument("section")
+@click.argument("option")
+@click.argument("value")
+def metalog_add_option(logfile, section, option, value):
+    """Add an option to a section of the metadata.log file.
+
+    E.g. using an 'iterations' arg for the option
+    iterations: 1-iter, 2-iter, 3-iter
+    where the iterations are in the <iterations.file>, one iteration per line
+    """
+    add_metalog_option(logfile, section, option, value)
+
+
+#
+# results subcommand
+#
+@click.group()
+@click.pass_context
+def results(ctx):
+    """Used for passing the context from the main cli"""
+    pass
+
+
+@results.command()
+@options.pbench_upload_user
+@options.pbench_server_prefix
+@options.pbench_show_server
+def move(user, prefix, show_server):
+    move_results(user, prefix, show_server)
 
 
 @click.group()
@@ -13,5 +57,5 @@ def main(ctx):
     return 0
 
 
-main.add_command(metalog.metalog)
-main.add_command(results.results)
+main.add_command(metalog)
+main.add_command(results)
