+++ pbench/test/agent/task/test_move_results.py
@@ -2,7 +2,7 @@ import datetime
 import responses
 
 from pbench.agent.config import AgentConfig
-from pbench.agent.task.move_results import move_results
+from pbench.cli.agent.commands.results import move_results
 from pbench.test.agent.task.common import MockDatetime, mock_agent_config
